Add experimental checkpointing
